+++ src/components/pages/inventory/InventoryItemDescription.tsx
@@ -33,15 +33,12 @@ export function InventoryItemDescription({
 }: InventoryItemDescriptionProps) {
   const hasItemInUse = !!itemInUse
 
-  const isProfileItem =
-    itemType === 'capas de perfil' || itemType === 'cards de jogador'
-
   return (
     <div className="flex-initial flex-col gap-6">
       {item?.id !== 0 && (
         <>
           <div className="items-center gap-2">
-            {item?.subtype && !isProfileItem && (
+            {item?.subtype && (
               <>
                 <span
                   className={twMerge(
@@ -66,14 +63,12 @@ export function InventoryItemDescription({
               {item.description}
             </p>
 
-            {!isProfileItem && (
-              <span className={twMerge('text-sm text-white', 'leading-none')}>
-                No jogo desde -{' '}
-                <span className="font-semibold leading-none">
-                  {formatDateToPtFormat(item.release_date)}
-                </span>
+            <span className={twMerge('text-sm text-white', 'leading-none')}>
+              No jogo desde -{' '}
+              <span className="font-semibold leading-none">
+                {formatDateToPtFormat(item.release_date)}
               </span>
-            )}
+            </span>
           </div>
         </>
       )}
@@ -92,8 +87,7 @@ export function InventoryItemDescription({
               }
             >
               <Button.Content className="text-sm font-semibold">
-                Remover{' '}
-                {isProfileItem ? 'Customizável' : removeSFromEnd(itemType)}
+                Remover {removeSFromEnd(itemType)}
               </Button.Content>
             </Button.Root>
           )}
@@ -105,8 +99,7 @@ export function InventoryItemDescription({
               onClick={() => handleUpdateItemInUse({ item_id: item.id })}
             >
               <Button.Content className="text-sm font-semibold">
-                {item.in_use ? 'Remover' : 'Ativar'}{' '}
-                {isProfileItem ? 'Customizável' : removeSFromEnd(itemType)}
+                {item.in_use ? 'Remover' : 'Ativar'} {removeSFromEnd(itemType)}
               </Button.Content>
             </Button.Root>
           )}