+++ CHANGELOG.md
@@ -59,6 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   
 ### Changed
 
+- Move customização de `cards` e `capas` para a página de `/conta` [#968](https://github.com/3C-gg/reload-frontend/issues/968).
 - Altera `Carousel` da loja para ficar igual ao novo design [#988](https://github.com/3C-gg/reload-frontend/issues/988).
 - Altera `initialSlide` no componente `CenteredCarousel`.
 - Altera projeto `Next` para ficar igual ao projeto antigo [#971](https://github.com/3C-gg/reload-frontend/issues/971).